Add import filter
Adds a new optional setting on a composition graph to filter imports for
inclusion into the graph.

Any WASM component import can be excluded for "forced" into the graph.

For example, an import of an interface that is defined statically in the
host linker should be excluded since it is not expected to be provided
as a dynamically link component in the graph.